requirements.txt
langchain
langchain_core
python_dotenv
langchain_community
streamlit
